Don't copy the keys if /var/yum/www/SIMP does not exist

+++ build/simp-gpgkeys.spec
@@ -46,52 +46,52 @@ cp GPGKEYS/* %{buildroot}/%{prefix}
 export PATH=/opt/puppetlabs/bin:$PATH
 
 # If we're a SIMP server, place the keys into the appropriate web directory
-
-dir='/var/www/yum/SIMP/GPGKEYS'
-if [ ! -d $dir ]; then
-  mkdir -p -m 0755 "${dir}"
-fi
-cp %{prefix}/RPM-GPG-KEY* "${dir}"
-
-# Get rid of any files that are present that aren't in the new directory.
-# Ensure that we don't have issues with operations in progress.
-old_key_list=`mktemp --suffix=.simp_gpgkeys`
-new_key_list=`mktemp --suffix=.simp_gpgkeys`
-
-find "${dir}" -name "RPM-GPG-KEY*" -maxdepth 1 -printf "%f\n" | sort -u > $old_key_list
-find "%{prefix}" -name "RPM-GPG-KEY*" -maxdepth 1 -printf "%f\n" | sort -u > $new_key_list
-
-for file in `comm -23 $old_key_list $new_key_list`; do
-  if [ -f "${dir}/${file}" ]; then
-    rm -f "${dir}/${file}"
+if [ -d '/var/www/yum/SIMP' ]; then
+  dir='/var/www/yum/SIMP/GPGKEYS'
+  if [ ! -d $dir ]; then
+    mkdir -p -m 0755 "${dir}"
   fi
-done
+  cp %{prefix}/RPM-GPG-KEY* "${dir}"
 
-if [ -f $old_key_list ]; then
-  rm -f $old_key_list
-fi
+  # Get rid of any files that are present that aren't in the new directory.
+  # Ensure that we don't have issues with operations in progress.
+  old_key_list=`mktemp --suffix=.simp_gpgkeys`
+  new_key_list=`mktemp --suffix=.simp_gpgkeys`
 
-if [ -f $new_key_list ]; then
-  rm -f $new_key_list
-fi
+  find "${dir}" -name "RPM-GPG-KEY*" -maxdepth 1 -printf "%f\n" | sort -u > $old_key_list
+  find "%{prefix}" -name "RPM-GPG-KEY*" -maxdepth 1 -printf "%f\n" | sort -u > $new_key_list
 
-# Link system GPG keys into SIMP repo
-if [ `facter operatingsystem` == 'CentOS' ]; then
-  search_string='.*CentOS-[[:digit:]]'
-elif [ `facter operatingsystem` == 'RedHat' ]; then
-  search_string='.*redhat.*release.*'
-else
-  search_string=''
-fi
-if [ -n "$search_string" ]; then
-  for file in `find /etc/pki/rpm-gpg/ -regextype posix-extended -regex ${search_string}`; do
-    cp ${file} ${dir}
+  for file in `comm -23 $old_key_list $new_key_list`; do
+    if [ -f "${dir}/${file}" ]; then
+      rm -f "${dir}/${file}"
+    fi
   done
-fi
 
-# Ensure GPG permissions
-chown -R root:48 ${dir}
-find ${dir} -type f -exec chmod 640 {} +
+  if [ -f $old_key_list ]; then
+    rm -f $old_key_list
+  fi
+
+  if [ -f $new_key_list ]; then
+    rm -f $new_key_list
+  fi
+
+  # Link system GPG keys into SIMP repo
+  if [ `facter operatingsystem` == 'CentOS' ]; then
+    search_string='.*CentOS-[[:digit:]]'
+  elif [ `facter operatingsystem` == 'RedHat' ]; then
+    search_string='.*redhat.*release.*'
+  else
+    search_string=''
+  fi
+  if [ -n "$search_string" ]; then
+    for file in `find /etc/pki/rpm-gpg/ -regextype posix-extended -regex ${search_string}`; do
+      cp ${file} ${dir}
+    done
+  fi
+  # Ensure GPG permissions
+  chown -R root:48 ${dir}
+  find ${dir} -type f -exec chmod 640 {} +
+fi
 
 %changelog
 * Thu Apr 22 2021 Jeanne Greulich <jeanne.greulich@gmail.com> - 3.1.2-0
